Implement a simple problem wrapper
This commit implements a simple, experimental "problem" error
definition.

It is designed to provide some context to users who are running the
application what they can do to help fix it.

It is based on the RFC at:

  https://tools.ietf.org/html/rfc7807

== Design Notes
=== Internal Package

This package is still experimental, but this application is designed to
be consumed by third parties.

Given this, documenting the errors arguably makes sense.

=== Usage

This library is used in a small way to remind users they need to add
claims.

Claim validation is now made a requirement, as the alternative is
"failing open" if the application does not detect any claims, and
"failing closed" is considered safer.