Allow multiple matching claim sets
The project for which this is being built (incman.site) is designed to
allow mutliple different people from within the same single sign on
source to review the site. Specifically, people from multiple different
organisations.

To do this, the application needs to allow a "at least 1 matching" type
authentication.

This commit implements that by supplying multiple oidc claims,
validating all of them, and just exiting if *none* pass.

== Design Notes
=== Interface

The initial way this was designed was in the form of:

oneOf:
  - {} // Claim Set

allOf:
  - {} // Claim Set

However, this design is implicitly possible if we just accept ${N} claim
sets, each of which can be valid.

Should users want one claim set that they desire to be valid, they
submit *only* that claim set.

The logic is simpler.