Make the books friendlier to read
Currently the design of the ePub that is being used to test this library
is such that it expects to be read on a screen reader.

That has a couple of notable properties, including:

- It takes up the full width of the limited screen real estate
- It doesn't include the meta properties mobile devices are expecting

This commit seeks to address that by modifying the templates in place to
include these additional references.

== Design Notes
=== Diiirty

The design of this, more generally, is not elegant. Rather, it is "what
is required" to carry this project on, with future work being done later
to tidy it up. Maybe. It depends on whether or not this project sees
success.