17 - Read correct cookie for redirect URL
Currently there is a problem in which, after users are authenticated,
they are redirected to a nonsense URL `(.well-known/TOKEN)`.

This is likely as a result of reading the authentication cookie as a
URL, and attempting to make sense of that.

This commit addresses this by reading from the originally intended
redirect cookie URL, rather than the token.

Fixes #17